% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChromBackendMemory.R
\name{ChromBackendMemory}
\alias{ChromBackendMemory}
\alias{backendInitialize,ChromBackendMemory-method}
\title{Improved in-memory Chromatographic data backend}
\usage{
ChromBackendMemory()

\S4method{backendInitialize}{ChromBackendMemory}(
  object,
  chromData = fillCoreChromVariables(data.frame()),
  peaksData = list(.EMPTY_PEAKS_DATA),
  ...
)
}
\arguments{
\item{object}{A \code{ChromBackendMemory} object.}

\item{chromData}{For \code{backendInitialize()} of a \code{ChromBackendMemory} backend,
a \code{data.frame} with the chromatographic data. If not provided
(or if empty), a default \code{data.frame} with the core chromatographic
variables will be created.}

\item{peaksData}{For \code{backendInitialize()} of a \code{ChromBackendMemory} backend,
a \code{list} of \code{data.frame} with the peaks data. If not provided (or if
empty), a default \code{list} of empty \code{data.frame} with the core peaks
variables will be created. The length of the list should match the number
of chromatograms in the \code{chromData} parameter.}

\item{...}{Additional parameters to be passed.}
}
\description{
\code{ChromBackendMemory}: This backend stores chromatographic data directly
in memory, making it ideal for small datasets or testing. It can be
initialized with a \code{data.frame} of chromatographic data via the \code{chromData}
parameter and a \code{list} of \code{data.frame} entries for peaks data using the
\code{peaksData} parameter. These data can be accessed with the \code{chromData()} and
\code{peaksData()} functions.
}
\author{
Philippine Louail
}
