% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChromBackendMzR.R
\name{ChromBackendMzR}
\alias{ChromBackendMzR}
\alias{[,ChromBackendMzR-method}
\alias{backendInitialize,ChromBackendMzR-method}
\title{Chromatographic data backend for reading mzML files.}
\usage{
ChromBackendMzR()

\S4method{backendInitialize}{ChromBackendMzR}(object, files = character(), BPPARAM = bpparam(), ...)
}
\arguments{
\item{object}{A \code{ChromBackendMzR} object.}

\item{files}{A character vector of file paths to mzML files.}

\item{BPPARAM}{Parallel setup configuration. See \code{\link[BiocParallel:register]{BiocParallel::bpparam()}}
for more information.}

\item{...}{Additional parameters to be passed.}
}
\description{
The \code{ChromBackendMzR} inherits all slots and methods from the base
\code{ChromBackendMemory} backend, providing additional functionality for
reading chromatographic data from mzML files.

Contrarily to the \code{ChromBackendMemory} backend, the \code{ChromBackendMzR} backend
should have the \emph{dataOrigin} chromatographic variables populated with the
file path of the mzML file from which the chromatographic data was read.

It should be noted that the \code{ChromBackendMzR} backend is read-only, and does
not support direct modification of chromatographic data. However, it does
support \code{peaksData} slot replacement, which will modify the \code{peaksData} slot but
not the local mzML files. This is the  signaled by the "inMemory" being set
to TRUE.

Lastly, implementing functionalities with the \code{ChromBackendMzR} backend
should be simplified as much as possible and reuse the methods already
implemented for \code{ChromBackendMemory} when possible.
}
\author{
Philippine Louail
}
