% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChromBackend-functions.R
\name{validChromData}
\alias{validChromData}
\title{Check core chromatogram variables for correct data types}
\usage{
validChromData(x = data.frame(), error = TRUE)
}
\arguments{
\item{x}{\code{data.frame} representing metadata of a \code{Chromatograms}}

\item{error}{\code{logical(1)} whether an error should be thrown (the default)
if one or more columns don't have the correct data type.}
}
\value{
If core variables have all the correct data type: an empty character.
If one or more core variables (columns) have the wrong data type the
function either throws an error (with \code{error = TRUE}) or returns a
\code{character} specifying which variables/columns don't have the correct
type (for \code{error = FALSE}).
}
\description{
\code{validChromData()} checks that columns, representing \emph{core} chromatogram
variables are of the correct data type.
}
