% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chromatograms.R
\name{Chromatograms}
\alias{Chromatograms}
\alias{Chromatograms-class}
\title{The Chromatograms class to manage and access chromatographic data}
\usage{
Chromatograms(backend = ChromBackendMemory(), processingQueue = list(), ...)
}
\arguments{
\item{backend}{\link{ChromBackend} object providing the raw data for the
\code{Chromatograms} object.}

\item{processingQueue}{\link{list} a list of processing steps to be applied to the
chromatograms data. Each element in the list is a function that
processes the chromatograms data. The processing steps are applied in
the order they are listed in the \code{processingQueue}.}

\item{...}{Additional arguments.}
}
\description{
The \code{Chromatograms} class encapsules chromatographic data and related metadata.
The chromatographic data is represented by a \emph{backend} extending the virtual
\link{ChromBackend} class which provides the raw data to the \code{Chromatograms}
object. Different backends and their properties are explained in the
\link{ChromBackend} class documentation.
}
\section{Data stored in a \code{Chromatograms} object}{


The \code{Chromatograms} object is a container for chromatographic data that includes
peaks data (\emph{retention time} and related intensity values,
also referred to as \emph{peaks data variables} in the context of \code{Chromatograms})
and metadata of individual chromatogram (so called
\emph{chromatograms variables}). While a core set of chromatograms variables (the
\code{coreChromatogramsVariables()}) and peaks data variables (the
\code{corePeaksVariables()}) are guaranteed to be provided by a \code{Chromatograms},
it is possible to add arbitrary additional chromatograms variables to a
\code{Chromatograms} object.

The \code{Chromatograms} object is designed to contain chromatographic data of a
(large) set of chromatograms. The data is organized \emph{linearly} and can be
thought of a list of chromatograms, i.e. each element in the \code{Chromatograms}
is one chomatogram.
}

\section{Creation of objects}{


\code{Chromatograms} objects can be created using the \code{Chromatograms()}
construction function.
}

\examples{
## create a Chromatograms object
chroms <- Chromatograms(backend = ChromBackendMemory())

}
