% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChromBackend.R
\docType{methods}
\name{ChromBackend}
\alias{ChromBackend}
\alias{backendInitialize,ChromBackend-method}
\alias{backendMerge,list-method}
\alias{backendMerge,ChromBackend-method}
\alias{dataOrigin,ChromBackend-method}
\alias{dataOrigin<-,ChromBackend-method}
\alias{dataStorage,ChromBackend-method}
\alias{dataStorage<-,ChromBackend-method}
\alias{intensity,ChromBackend-method}
\alias{intensity<-,ChromBackend-method}
\alias{isEmpty,ChromBackend-method}
\alias{isReadOnly,ChromBackend-method}
\alias{length,ChromBackend-method}
\alias{msLevel,ChromBackend-method}
\alias{msLevel<-,ChromBackend-method}
\title{Chromatographic data backends}
\usage{
\S4method{backendInitialize}{ChromBackend}(object, ...)

\S4method{backendMerge}{list}(object, ...)

\S4method{backendMerge}{ChromBackend}(object, ...)

\S4method{dataOrigin}{ChromBackend}(object)

\S4method{dataOrigin}{ChromBackend}(object) <- value

\S4method{dataStorage}{ChromBackend}(object)

\S4method{dataStorage}{ChromBackend}(object) <- value

\S4method{intensity}{ChromBackend}(object)

\S4method{intensity}{ChromBackend}(object) <- value

\S4method{isEmpty}{ChromBackend}(x)

\S4method{isReadOnly}{ChromBackend}(object)

\S4method{length}{ChromBackend}(x)

\S4method{msLevel}{ChromBackend}(object)

\S4method{msLevel}{ChromBackend}(object) <- value
}
\description{
\code{ChromBackend} is a virtual class that defines what different backends need
to provide to be used by the \code{Chromatograms} package and classes.
}
\author{
Johannes Rainer, Sebastian Gibb, Laurent Gatto
}
