% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/AllGenerics.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{backendInitialize}
\alias{backendMerge}
\alias{dataOrigin}
\alias{dataOrigin<-}
\alias{dataStorage}
\alias{dataStorage<-}
\alias{isReadOnly}
\title{Internal page for hidden aliases}
\usage{
backendInitialize(object, ...)

backendMerge(object, ...)

dataOrigin(object, ...)

dataOrigin(object) <- value

dataStorage(object, ...)

dataStorage(object) <- value

isReadOnly(object, ...)
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
