% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChromBackend-functions.R
\name{validPeaksData}
\alias{validPeaksData}
\title{\code{validPeaksData()} checks that the names of the input peaksData list,
representing \emph{core} peaks variables are of the correct data type.}
\usage{
validPeaksData(x = list())
}
\arguments{
\item{x}{\code{list} representing the peaks data of a \code{Chromatograms}}
}
\description{
\code{validPeaksData()} checks that the names of the input peaksData list,
representing \emph{core} peaks variables are of the correct data type.
}
