% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/AllGenerics.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{backendInitialize}
\alias{backendMerge}
\alias{chromData}
\alias{chromData<-}
\alias{chromNames}
\alias{chromNames<-}
\alias{chromVariables}
\alias{isReadOnly}
\alias{mzMax}
\alias{mzMax<-}
\alias{mzMin}
\alias{mzMin<-}
\alias{pairs}
\alias{pairs<-}
\alias{precursorMzMin}
\alias{precursorMzMin<-}
\alias{precursorMzMax}
\alias{precursorMzMax<-}
\alias{productMzMax}
\alias{productMzMax<-}
\alias{productMzMin}
\alias{productMzMin<-}
\alias{selectChromVariables}
\title{Internal page for hidden aliases}
\usage{
backendInitialize(object, ...)

backendMerge(object, ...)

chromData(object, ...)

chromData(object) <- value

chromNames(object, ...)

chromNames(object) <- value

chromVariables(object, ...)

isReadOnly(object, ...)

mzMax(object, ...)

mzMax(object) <- value

mzMin(object, ...)

mzMin(object) <- value

pairs(object, ...)

pairs(object) <- value

precursorMzMin(object, ...)

precursorMzMin(object) <- value

precursorMzMax(object, ...)

precursorMzMax(object) <- value

productMzMax(object, ...)

productMzMax(object) <- value

productMzMin(object, ...)

productMzMin(object) <- value

selectChromVariables(object, ...)
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
