% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_aliases.R, R/AllGenerics.R,
%   R/ChromBackendDataFrame.R
\docType{methods}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{backendInitialize}
\alias{backendMerge}
\alias{chromData}
\alias{chromData<-}
\alias{chromNames}
\alias{chromNames<-}
\alias{chromVariables}
\alias{isReadOnly}
\alias{mzMax}
\alias{mzMax<-}
\alias{mzMin}
\alias{mzMin<-}
\alias{precursorMzMin}
\alias{precursorMzMin<-}
\alias{precursorMzMax}
\alias{precursorMzMax<-}
\alias{productMzMax}
\alias{productMzMax<-}
\alias{productMzMin}
\alias{productMzMin<-}
\alias{selectChromVariables}
\alias{show,ChromBackendDataFrame-method}
\alias{backendInitialize,ChromBackendDataFrame-method}
\alias{backendMerge,ChromBackendDataFrame-method}
\alias{as.list,ChromBackendDataFrame-method}
\alias{chromData,ChromBackendDataFrame-method}
\alias{chromData<-,ChromBackendDataFrame-method}
\alias{chromNames,ChromBackendDataFrame-method}
\alias{chromNames<-,ChromBackendDataFrame-method}
\alias{chromVariables,ChromBackendDataFrame-method}
\alias{dataOrigin,ChromBackendDataFrame-method}
\alias{dataOrigin<-,ChromBackendDataFrame-method}
\alias{dataStorage,ChromBackendDataFrame-method}
\alias{dataStorage<-,ChromBackendDataFrame-method}
\alias{intensity,ChromBackendDataFrame-method}
\alias{intensity<-,ChromBackendDataFrame-method}
\alias{isEmpty,ChromBackendDataFrame-method}
\alias{length,ChromBackendDataFrame-method}
\alias{lengths,ChromBackendDataFrame-method}
\alias{msLevel,ChromBackendDataFrame-method}
\alias{msLevel<-,ChromBackendDataFrame-method}
\alias{mz,ChromBackendDataFrame-method}
\alias{mz<-,ChromBackendDataFrame-method}
\alias{mzMax,ChromBackendDataFrame-method}
\alias{mzMax<-,ChromBackendDataFrame-method}
\alias{mzMin,ChromBackendDataFrame-method}
\alias{mzMin<-,ChromBackendDataFrame-method}
\alias{precursorMz,ChromBackendDataFrame-method}
\alias{precursorMz<-,ChromBackendDataFrame-method}
\alias{precursorMzMax,ChromBackendDataFrame-method}
\alias{precursorMzMax<-,ChromBackendDataFrame-method}
\alias{precursorMzMin,ChromBackendDataFrame-method}
\alias{precursorMzMin<-,ChromBackendDataFrame-method}
\alias{productMz,ChromBackendDataFrame-method}
\alias{productMz<-,ChromBackendDataFrame-method}
\alias{productMzMax,ChromBackendDataFrame-method}
\alias{productMzMax<-,ChromBackendDataFrame-method}
\alias{productMzMin,ChromBackendDataFrame-method}
\alias{productMzMin<-,ChromBackendDataFrame-method}
\alias{rtime,ChromBackendDataFrame-method}
\alias{rtime<-,ChromBackendDataFrame-method}
\alias{selectChromVariables,ChromBackendDataFrame-method}
\alias{$,ChromBackendDataFrame-method}
\alias{$<-,ChromBackendDataFrame-method}
\alias{[,ChromBackendDataFrame-method}
\alias{filterDataOrigin,ChromBackendDataFrame-method}
\alias{filterDataStorage,ChromBackendDataFrame-method}
\alias{filterMsLevel,ChromBackendDataFrame-method}
\alias{filterMz,ChromBackendDataFrame-method}
\alias{filterPrecursorMz,ChromBackendDataFrame-method}
\alias{filterProductMz,ChromBackendDataFrame-method}
\title{Internal page for hidden aliases}
\usage{
backendInitialize(object, ...)

backendMerge(object, ...)

chromData(object, ...)

chromData(object) <- value

chromNames(object, ...)

chromNames(object) <- value

chromVariables(object, ...)

isReadOnly(object, ...)

mzMax(object, ...)

mzMax(object) <- value

mzMin(object, ...)

mzMin(object) <- value

precursorMzMin(object, ...)

precursorMzMin(object) <- value

precursorMzMax(object, ...)

precursorMzMax(object) <- value

productMzMax(object, ...)

productMzMax(object) <- value

productMzMin(object, ...)

productMzMin(object) <- value

selectChromVariables(object, ...)

\S4method{show}{ChromBackendDataFrame}(object)

\S4method{backendInitialize}{ChromBackendDataFrame}(object, chromData, ...)

\S4method{backendMerge}{ChromBackendDataFrame}(object, ...)

\S4method{as.list}{ChromBackendDataFrame}(x, ...)

\S4method{chromData}{ChromBackendDataFrame}(object,
  columns = chromVariables(object))

\S4method{chromData}{ChromBackendDataFrame}(object) <- value

\S4method{chromNames}{ChromBackendDataFrame}(object)

\S4method{chromNames}{ChromBackendDataFrame}(object) <- value

\S4method{chromVariables}{ChromBackendDataFrame}(object)

\S4method{dataOrigin}{ChromBackendDataFrame}(object)

\S4method{dataOrigin}{ChromBackendDataFrame}(object) <- value

\S4method{dataStorage}{ChromBackendDataFrame}(object)

\S4method{dataStorage}{ChromBackendDataFrame}(object) <- value

\S4method{intensity}{ChromBackendDataFrame}(object)

\S4method{intensity}{ChromBackendDataFrame}(object) <- value

\S4method{isEmpty}{ChromBackendDataFrame}(x)

\S4method{length}{ChromBackendDataFrame}(x)

\S4method{lengths}{ChromBackendDataFrame}(x)

\S4method{msLevel}{ChromBackendDataFrame}(object, ...)

\S4method{msLevel}{ChromBackendDataFrame}(object) <- value

\S4method{mz}{ChromBackendDataFrame}(object, ...)

\S4method{mz}{ChromBackendDataFrame}(object) <- value

\S4method{mzMax}{ChromBackendDataFrame}(object, ...)

\S4method{mzMax}{ChromBackendDataFrame}(object) <- value

\S4method{mzMin}{ChromBackendDataFrame}(object, ...)

\S4method{mzMin}{ChromBackendDataFrame}(object) <- value

\S4method{precursorMz}{ChromBackendDataFrame}(object)

\S4method{precursorMz}{ChromBackendDataFrame}(object) <- value

\S4method{precursorMzMax}{ChromBackendDataFrame}(object)

\S4method{precursorMzMax}{ChromBackendDataFrame}(object) <- value

\S4method{precursorMzMin}{ChromBackendDataFrame}(object)

\S4method{precursorMzMin}{ChromBackendDataFrame}(object) <- value

\S4method{productMz}{ChromBackendDataFrame}(object)

\S4method{productMz}{ChromBackendDataFrame}(object) <- value

\S4method{productMzMax}{ChromBackendDataFrame}(object)

\S4method{productMzMax}{ChromBackendDataFrame}(object) <- value

\S4method{productMzMin}{ChromBackendDataFrame}(object)

\S4method{productMzMin}{ChromBackendDataFrame}(object) <- value

\S4method{rtime}{ChromBackendDataFrame}(object)

\S4method{rtime}{ChromBackendDataFrame}(object) <- value

\S4method{selectChromVariables}{ChromBackendDataFrame}(object,
  chromVariables = chromVariables(object))

\S4method{$}{ChromBackendDataFrame}(x, name)

\S4method{$}{ChromBackendDataFrame}(x, name) <- value

\S4method{[}{ChromBackendDataFrame}(x, i, j, ..., drop = FALSE)

\S4method{filterDataOrigin}{ChromBackendDataFrame}(object,
  dataOrigin = character(), ...)

\S4method{filterDataStorage}{ChromBackendDataFrame}(object,
  dataStorage = character(), ...)

\S4method{filterMsLevel}{ChromBackendDataFrame}(object,
  msLevel = integer(), ...)

\S4method{filterMz}{ChromBackendDataFrame}(object, mz = numeric(), ...)

\S4method{filterPrecursorMz}{ChromBackendDataFrame}(object,
  mz = numeric(), ...)

\S4method{filterProductMz}{ChromBackendDataFrame}(object, mz = numeric(),
  ...)
}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
