% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chromatograms-peaksData.R
\name{peaksData}
\alias{peaksData}
\alias{peaksVariables}
\alias{filterPeaksData}
\alias{applyProcessing}
\alias{peaksData,Chromatograms-method}
\alias{peaksData<-,Chromatograms-method}
\alias{peaksVariables,Chromatograms-method}
\alias{rtime,Chromatograms-method}
\alias{rtime<-,Chromatograms-method}
\alias{intensity,Chromatograms-method}
\alias{intensity<-,Chromatograms-method}
\alias{filterPeaksData,Chromatograms-method}
\alias{applyProcessing,Chromatograms-method}
\alias{addProcessing,Chromatograms-method}
\title{Chromatographic peaks data.}
\usage{
\S4method{peaksData}{Chromatograms}(
  object,
  columns = peaksVariables(object),
  f = processingChunkFactor(object),
  BPPARAM = bpparam(),
  drop = FALSE,
  ...
)

\S4method{peaksData}{Chromatograms}(object) <- value

\S4method{peaksVariables}{Chromatograms}(object, ...)

\S4method{rtime}{Chromatograms}(object, ...)

\S4method{rtime}{Chromatograms}(object) <- value

\S4method{intensity}{Chromatograms}(object, ...)

\S4method{intensity}{Chromatograms}(object) <- value

\S4method{filterPeaksData}{Chromatograms}(
  object,
  variables = character(),
  ranges = numeric(),
  match = c("any", "all"),
  keep = TRUE
)

\S4method{applyProcessing}{Chromatograms}(
  object,
  f = processingChunkFactor(object),
  BPPARAM = bpparam(),
  ...
)

\S4method{addProcessing}{Chromatograms}(object, FUN, ...)
}
\arguments{
\item{object}{A \link{Chromatograms} object.}

\item{columns}{For \code{peaksData()} accessor: optional \code{character} with column
names (peaks variables) that should be included in the
returned \code{list} of \code{data.frame}. By default, all columns are returned.}

\item{f}{\code{factor} defining the grouping to split the \code{Chromatograms} object.}

\item{BPPARAM}{Parallel setup configuration. See \code{\link[BiocParallel:register]{BiocParallel::bpparam()}}
for more information.}

\item{drop}{\code{logical(1)} default to \code{FALSE}. If \code{TRUE}, and one column is
called by the user, the method should  return a vector (or list of
vector for \code{peaksData()}) of the single column requested.}

\item{...}{Additional arguments passed to the method.}

\item{value}{For \code{rtime()} and \code{intensity()}: \code{numeric} vector with the
values to replace the current values.}

\item{variables}{For \code{filterPeaksData()}: \code{character} vector with the names
of the chromatogram variables to filter for. The list of available
chromatogram variables can be obtained with \code{peaksVariables()}.}

\item{ranges}{For \code{filterPeaksData()} : a \code{numeric}
vector of paired values (upper and lower boundary) that define the
ranges to filter the \code{object}. These paired values need to be in the
same order as the \code{variables} parameter (see below).}

\item{match}{For \code{filterPeaksData()} : \code{character(1) }
defining whether the condition has to match for all provided
\code{ranges} (\code{match = "all"}; the default), or for any of them
(\code{match = "any"}) for chromatogram data to be retained.}

\item{keep}{For \code{filterPeaksData()}: \code{logical(1)}
defining whether to keep (\code{keep = TRUE}) or remove (\code{keep = FALSE})
the chromatogram data that match the condition.}

\item{FUN}{FOr \code{addProcessing()} A function to be added to the
\code{Chromatograms} object's processing queue.}
}
\description{
As explained in the \code{\link{Chromatograms}} class documentation, the \code{Chromatograms}
object is a container for chromatogram data that includes chromatographic
peaks data (\emph{retention time} and related intensity values, also referred to
as \emph{peaks data variables} in the context of \code{Chromatograms}) and metadata of
individual chromatograms (so called \emph{chromatograms variables}).

The \emph{peaks data variables} information can be accessed using the
\code{peaksData()} function. it is also possible to access specific
peaks variables using \code{$}.

\code{peaks} can be accessed, replaced but also filtered/subsetted. Refer to
the sections below for more details.
}
\section{Filter peaks variables}{


Functions that filter \code{Chromatograms} based on peaks variables
(i.e, \code{peaksData} ) will remove peaks data that do not meet the
specified conditions. This means that if a chromatogram is filtered out, its
corresponding only the peaks variables pairs in the \code{peaksData} will be
removed from the object. Moreover, to keep things effcient and all peaks data
can get quite large, a processing queue is put in place, therefore the actual
backend data stays the same until the processing queue is applied (i.e. by
running \code{applyProcessing()}). When calling \code{peaksData()} the processing queue
is applied in the output but the backend data is not replaced.

The available functions to filter chromatogram data are:
\itemize{
\item \code{filterPeaksData()}: Filters numerical peaks data variables
based on the provided numerical \code{ranges}. The method returns the same input a
\code{Chromatograms} object but the filtering step is added to the processing
queue, which will then be applied if the user later request for access to
the \code{peaksData}. This function will \emph{not} results in an object with
fewer chromatograms than the original, however the peaks data ("rtime" and
"intensity' data pair, for example) will be removed from the object.
}
}

\seealso{
\link{Chromatograms} for a general description of the \code{Chromatograms}
object, and \link{chromData} for accessing,substituting and filtering
chromatographic variables.
}
\author{
Philippine Louail
}
