% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chromatograms-plot.R
\name{plotChromatograms}
\alias{plotChromatograms}
\alias{plotChromatograms,Chromatograms-method}
\alias{plotChromatogramsOverlay,Chromatograms-method}
\title{Plot chromatograms}
\usage{
\S4method{plotChromatograms}{Chromatograms}(
  object,
  xlab = "rtime (s)",
  ylab = "intensity",
  xlim = numeric(),
  ylim = numeric(),
  main = character(),
  col = "#00000080",
  pch = 20,
  cex = 1.5,
  lwd = 0.5,
  bs = 16,
  ...
)

\S4method{plotChromatogramsOverlay}{Chromatograms}(
  object,
  xlab = "rtime (s)",
  ylab = "intensity",
  xlim = numeric(),
  ylim = numeric(),
  main = paste(length(object), "chromatograms"),
  col = "#00000080",
  pch = 20,
  cex = 1.5,
  lwd = 0.5,
  bs = 16,
  ...
)
}
\arguments{
\item{object}{A \link{Chromatograms} object.}

\item{xlab}{\code{character(1)} with the label for the x-axis (by default
\code{xlab = "rtime (s)"}).}

\item{ylab}{\code{character(1)} with the label for the y-axis (by default
\code{ylab = "intensity"}).}

\item{xlim}{\code{numeric(2)} defining the x-axis limits. The range of m/z values
are used by default.}

\item{ylim}{\code{numeric(2)} defining the y-axis limits. The range of intensity
values are used by default.}

\item{main}{\code{character(1)} with the title for the plot. By default the
spectrum's MS level and retention time (in seconds) is used.}

\item{col}{color to be used to draw the peaks. Should be either of length 1,
or equal to the number of chromatograms (to plot each chromatograms
in a different color) or be a \code{list} with colors for each individual
peak in each spectrum.}

\item{pch}{\code{integer(1)} or \code{character(1)} specifying the plotting symbol
(see \code{\link[=plot.default]{plot.default()}}).}

\item{cex}{\code{numeric(1)} specifying the size of the plotting symbol (see
\code{\link[=plot.default]{plot.default()}}).}

\item{lwd}{\code{numeric(1)} specifying the line width (see \code{\link[=plot.default]{plot.default()}}).}

\item{...}{Additional arguments to be passed to \code{\link[=plot.default]{plot.default()}}.}

\item{frame.plot}{\code{logical(1)} whether a box should be drawn around the
plotting area.}
}
\value{
These functions create a plot.

Refer to the individual function description for information on the
return value.
}
\description{
\code{\link[=Chromatograms]{Chromatograms()}} can be plotted with the following functions:

The \code{plotChromatograms()}: plots each chromatogram in its separate plot by
splitting the plot area into as many panels as there are spectra.
}
\examples{

## Create a Chromatograms object
cdata <- data.frame(
    msLevel = c(1L, 1L, 1L),
    mz = c(112.2, 123.3, 134.4),
    chromIndex = c(1L, 2L, 3L)
)
pdata <- list(
    data.frame(
        rtime = c(12.4, 12.8, 13.2, 14.6),
        intensity = c(123.3, 153.6, 2354.3, 243.4)
    ),
    data.frame(
        rtime = c(45.1, 46.2),
        intensity = c(100, 80.1)
    ),
    data.frame(
        rtime = c(12.4, 12.8, 13.2, 14.6),
        intensity = c(123.3, 153.6, 2354.3, 243.4)
    )
)
chr <- backendInitialize(ChromBackendMemory(),
    chromData = cdata,
    peaksData = pdata
) |> Chromatograms()

## Plot one chromatogram
plotChromatograms(chr[1])

## Plot the full Chromatograms object
plotChromatograms(chr)

## Define a color for each peak in each chromatogram
plotChromatograms(chr[1:2], col = c("green", "blue"))

## Overlay all chromatograms
plotChromatogramsOverlay(chr[1:2], col = c("green", "blue"))

}
\author{
Philippine Louail, Johannes Rainer.
}
